renamed selector.ts -> selectors.ts
