+++ src/store/slices/offers/selectors.ts
